+++ src/mc_classifier_pipeline/__init__.py
@@ -1,0 +1,5 @@
+"""MC Classifier Pipeline package."""
+
+from . import doc_retriever
+
+__all__ = ["doc_retriever"]